Front parcialmente implementado

+++ app/Http/Controllers/ClientController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreClientRequest;
 use App\Http\Requests\UpdateClientRequest;
 use App\Models\Client;
+use Illuminate\Support\Facades\Gate;
 
 class ClientController extends Controller
 {
@@ -15,6 +16,7 @@ class ClientController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Client::class);
         return view('clients.index', ['clients' => Client::all()]);
     }
 
@@ -36,6 +38,7 @@ class ClientController extends Controller
      */
     public function store(StoreClientRequest $request)
     {
+        Gate::authorize('create', Client::class);
         $validated = $request->validated();
         $client = Client::create($validated);
         return redirect()->route('client.index');
@@ -49,6 +52,7 @@ class ClientController extends Controller
      */
     public function show(Client $client)
     {
+        Gate::authorize('view', $client);
         return view('clients.show', ['client' => $client]);
     }
 
@@ -72,6 +76,7 @@ class ClientController extends Controller
      */
     public function update(UpdateClientRequest $request, Client $client)
     {
+        Gate::authorize('update', $client);
         $validated = $request->validated();
         $client->update($validated);
         return redirect()->route('client.index');
@@ -85,7 +90,21 @@ class ClientController extends Controller
      */
     public function destroy(Client $client)
     {
+        Gate::authorize('delete', $client);
         $client->delete();
+        return redirect()->route('client.index');
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Models\Client  $client
+     * @return \Illuminate\Http\Response
+     */
+    public function destroyAll()
+    {
+        Gate::authorize('deleteAll', Client::class);
+        Client::truncate();
         return redirect()->route('client.index');
     }
 }
